Fixed bug in disc numbering and writing album art to FLAC files.
There was an issue in parsing the iTunes XML file to obatin the disc numbers of an album that was resolved.
In the FLAC tagging program to add album art, there was an issue with code that 'allowed' the user to add mutilple images at one time. This code did NOT work correctly and has been removed until I have time to fix the issue.