Major update to functionality of code
With this update comes a rewrite of the majority
of the code, moving the main iTunes_Music_Converter
code to a python class, and adding the mutagen
package for file tagging. The ability to convert
multiple files at a time has also been added through
the threading module; likely is more efficient at
converting than copying this way. Some work still
must be done on the threading options. Also created
setup.py for easier install. Will install the
package, the command-line script, add the script
path to the user's path, and install the Applescript
into the ~/Library/iTunes/Scripts/ folder.

v0.2.1 20180202-2230 CDT