Adds dialog for status of conversion
This update adds a tkinter window with progress bars for the
status of the converstion. The window consists of an overall
status bar and then n status bars for the number of processes
that are allowed; i.e., if you have 8 CPUs, there will be
8 thread progress bars and an overall progress, so 9 bars
in total.

To do:
  1 - Fix code that does NOT realize file exists already
  2 - Clean up code and comment

v0.3.0 20190103 1030 CDT