Fixes progress bar & coverart bugs
This update fixes issues with the progress bars and
getting cover art.

For the progress bars, I realized there were multiple different
date/time strings that were matching the regex pattern set for
get the track duration and encoding time. This, in some cases,
caused the duration of the track to be enormous, which ment
the progress bar would not update correctly. This has been
fixed.

There was also an issue with years in the release data from
musicbrainz not parsing correctly, so added some more checks
in the utils.get_album_art functions.

Added some more threading locks (namely to the _get_cover method)
to ensure that threads aren't interfering with each other.

Also changed so that GUI does NOT close automatically, user
must exit out

v0.3.3 20190105 1334 CDT