Fixes minor bugs adds gui option
This update fixes some minor bugs in copying files when
file is already MP3; MP3 files are NOT re-encoded to help
safe some quality.

Other changes is a --gui option is add to the command line
program and gui keyword to the class __init__ funciton.
This will display the GUI progress bars. Still must
revert code, but the verbose keyword will give the old style
output to the command line

v0.3.4 20190105 1438 CDT